remove my_num_vars i8
testing this with wrf, and running out of memory when pushing
my_num_vars to i8, so leaving it out of this pull request. May
need to revisit.

+++ assimilation_code/modules/utilities/ensemble_manager_mod.f90
@@ -57,7 +57,7 @@ type ensemble_type
    !!!private
    integer(i8)                  :: num_vars
    integer                      :: num_copies, my_num_copies
-   integer(i8)                  :: my_num_vars
+   integer                      :: my_num_vars
    integer,        allocatable  :: my_copies(:)
    integer(i8),    allocatable  :: my_vars(:)
    ! Storage in next line is to be used when each pe has all copies of subset of vars
@@ -672,7 +672,7 @@ function get_my_num_vars(ens_handle)
 ! Returns the number of vars stored on my processor when copy complete.
 ! Same as num_vars if single process is in use.
 
-integer(i8)                      :: get_my_num_vars
+integer                          :: get_my_num_vars
 type (ensemble_type), intent(in) :: ens_handle
 
 get_my_num_vars = ens_handle%my_num_vars
@@ -712,7 +712,7 @@ subroutine set_up_ens_distribution(ens_handle)
 
 type (ensemble_type),  intent(inout)  :: ens_handle
 
-integer(i8) :: num_per_pe_below, num_left_over
+integer :: num_per_pe_below, num_left_over
 integer :: i
 
 ! Option 1: Maximum separation for both vars and copies
@@ -825,7 +825,7 @@ function get_max_num_vars(ens_handle, num_vars)
 
 type (ensemble_type), intent(in)  :: ens_handle
 integer(i8), intent(in) :: num_vars
-integer(i8)             :: get_max_num_vars
+integer                 :: get_max_num_vars
 !!!integer, intent(in) :: distribution_type
 
 !could this be instead:
